Cleaned up some code slightly and added a new welcoming message. I also added a charInput function.
Cleaned up some code slightly and added a new welcoming message. I also added a charInput function, which gets input from the user in the form of a character. This code will be used multiple times and is my first  function in C which returns a certain value.

Functions Added:
- charInput(char inputText[]) 
      returns userInput (string)